changed quantity from int to float considering the conversion

+++ src/Entity/Food.php
@@ -16,7 +16,7 @@ class Food
         private int $id,
         private string $name,
         private string $type,
-        private int $quantity,
+        private float $quantity,
         private string $unit = self::UNIT_GRAMS
     )
     {
